Canonicalize Rocks include paths and output them in `build.rs`
Our RocksDB wrapper crate needs to be able to locate the RocksDB include
files in order to compile its own C++ wrapper code.  This change uses a
feature of cargo that lets us print out the include paths used in
`build.rs`, and access those in an environment variable in the
`build.rs` dependent crates.