Implement flushing scalez-stor database to S3 (#839) (#927)
* Implement initial sync of RocksDB to S3 object storage

Summary
===
This task is part of epic #712, about scalez-stor database loading from
and persisting to S3.  This is just the initial implementation of the
persisting to S3; it doesn't contain any implementation for loading from
S3.  Also the persisting implementation is very naive.  There are
already several enhancements written up:

* #919 - Use multi-part upload for faster large file uploads
* #926 - Graceful shutdown of sync task to avoid transient error log
  output in tests
* #929 - Optimize how multiple commit operations interact to reduce
  total number of snapshots pushed to object storage
* #931 - Detect when the database hasn't changed since the last
  operation

Still, this feature by itself was a lot of work, more than I expected.

Details
===

Domovoi
---

I added quite a bit to `domovoi`.  Especially the `ObjectStorageBucket`
trait.  Most of the ops I added are obvious like `tag_objects`.  I also
added `smart_upload` and `smart_download` (although only the upload part
is implemented in this PR).  These sync a local file system to a remote
object store.

In order to control concurrency, I added a `Parallelism` type which is a
wrapper around a Tokio semaphore, and allows the caller of `domovoi` to
control the scope at which this limit applies.  Otherwise we'd have
unbounded parallelism which is a sure way to run out of memory.

Roxide (new name for assurio-rocks)
---
I thought this already had all I need, but I actually needed to do some
more work here.  I added a Flush operation that flushes all CFs, and
enabled atomic flushs so that such a flush operation actually preserves
data integrity.  I also enhanced the existing Checkpoint operation to
expose information about which SST files belong to which column
families, which will be important in the future because we won't want to
sync ALL CFs to S3 all of the time.

Finally I exposed `GetLatestSequenceNumber` which is very helpful for
uniquely identifying a particular point in time in the database, when
taking snapshots.

Scalez-KV
---

Here I added a new module `objstore` and struct `ObjectStorage`.  This
is intended to be a singleton which maintains a background worker task
and a queue of requests, running them one at a time.

It supports two operations:

* Sync, which is something that will happen automatically every minute or
so
* Snapshot, which depends on Sync but then uploads all database files
  and tags the objects so we can identify them as a snapshot and
  retrieve them later

There's some complexity here because the sync operation could take a few
seconds, or it could take several minutes, depending on load and S3
speed.  We don't ever want to do two sync operations at the same time,
because that's much more wasteful than saturating bandwidth to S3 on a
single sync.

I also don't want callers to have to block until their sync operation
happens.  So the API here is a bit more complex than you might think it
needs to be.

I also added checkpoint support here, which was not exposed before.  The
checkpoint code is mostly a wrapper around roxide with some convenience
functions that delete the checkpoint when the `Checkpoint` object is
dropped.

Scalez-Stor
---
This is where the rubber meets the road.  The already existing
`CommitTransaction` RPC calls are now modified to upload a snapshot of
the corresponding local RocksDB database to S3.  That is going to
increase the latency of the commit operation a lot, and I expect the
current naive implementation will be unacceptably slow.  However there
are a lot of opportunities for optimization, we'll see how big of a deal
the performance is here.  I'm also worried about S3 cost in production
as every snapshot involves a few PUT operations at least, but again
we'll learn from this version and iterate.

I also introduced a `ObjectStoreSyncer` type which handles setting up
the automatic background syncs implemented in scalez-kv, and also
handles taking snapshots.  The API around snapshots is a little weird
because I had a different idea about how snapshots would be done in
gRPC, and we may end up making that change in the commit API in the
future.

Scalez-Stor-Test
---

This crate is used for writing integration tests using scalez-stor.  It
got a big refactor

Closes #839