Add Sentry.io instrumentation to data plane (#1035)
* Add Sentry.io instrumentation to data plane

Summary
===

This adds a feature `sentry` to `scalez-kv`, `scalez-stor`, and `s0`.
when enabled, it installs the Sentry panic hook when the CLI starts, and
it also reports RocksDB background errors to Sentry.

Closes #789

Details
===

In RocksDB most errors are reported as failures to RocksDB API calls.
For now these will not be reported to Sentry, although they will
eventually appear there either as a `s0` CLI error or as an `elastio`
CLI error.  In the future we'll have deeper Sentry integration but for
now this is enough....except for background errors.

Background errors are errors that happen in RocksDB code that is running
in a worker thread, not directly associated with some API call.
Examples of this are flushing memtables to disk, and compacting.  If
these fail (maybe the disk is out of space for instance), the database
fails into a read-only mode, which is catastrophic and needs to be
reported.

There is an Events API in Rocks, but before this work the
BackgroundError event was not exposed.  So it had to be added to the
Rust RocksDB API, and then integrated into scalez.  In this
implementation, the background error handler is ALWAYS installed, but if
`sentry` is enabled the error is also reported to Sentry.  If not, it's
only logged and a metric incremented.

* Allow initializing sentry with a `ClientOptions` struct if `sentry` is enabled

* Forgot to actually report CLI errors to Sentry.

* Build `s0` with the `sentry` feature enabled for packaging