Update monorepo crates to Tokio 1.0 (#1402)
* Update monorepo crates to Tokio 1.0

Summary
===

This sounds like a relatively simple operation, update all `tokio` deps
from `0.2` to `1.0`.  How hard can be it?

It's a massive change, because it brings a number of transitive
dependency changes along with it, and those transitive dependencies must
all have been updated to themselves used tokio 1.0 and its dependencies.

A summary of the crates that need to be upgraded:

* `tokio`
* `tower`
* `tonic`
* `prost`
* `bytes`
* `rand`
* `quick-check`

Not only does this require to update alot of `Cargo.toml` dependencies,
but there are breaking API changes in `prost`, `rand`, and especially
`tokio`.  This has required me to actually modify some non-trivial code.

Key Tokio Changes
===

The most disruptive change for our code base is caused by the fact that
Tokio channel receiver types no longer implement `Stream`.  There are
complicated reasons for this.  A `Stream` wrapper is provided in a new
trait, `tokio-stream`, but only for some of the receiver types
(`broadcast` not included).

A related disruptive change is that the channel types are now `!Unpin`,
which broke a couple of places.

Another change is that `tokio::time::delay_for` is now
`tokio::time::sleep`.

Other Changes
===
`rand::get_range` now takes a `Range` and not separate lower and upper
boundary parameters.

A really shitty change is that the `binary-ids` crate generates a
`prost::Message` implementation that will now panic if a binary ID type
is deserialzied with the wrong length.  That is unlikely to happen as
long as we're using Rust clients and Rust servers, but once we start
interoperating with other languages that is more likely to cause a
problem.  This doesn't cause UB or expose a security vuln, but it's
really shitty ergonomics that I hope can be fixed in the future.

Also Prometheus integer metrics are now `u64` instead of `i64`.  That's
a welcome change but it did break all tests that were making assertions
about metrics values.

In `scalez-stor-client` I had `actix-web` as a `dev-dependencies` as an
easy way to spin up an HTTP server to test the HTTP download component.
But `actix-web` is still on the old Tokio (there's a beta version out
using 1.0 but all of the other actix deps still haven't been updated),
so I rewrote the HTTP server test to use hyper instead.

Dependencies
===
I change `dynomite` from using the release 0.10 to the latest master
(pinned to a specific rev for stability) because it's only very recently
been updated to tokio 1.0 and hasn't published a release yet.

I had to fork `rusty_ulid` and switch to that fork to start using the
new `rand` 0.8.  I've submitted that change upstream; once it's released
I'll be able to remove this fork.

Fix clippy lint in kesha tests

Remove rusty_ulid fork since the change was merged upstream

Move most dependencies to itertools 0.10

Fix some compile errors in test-data-generator

Fix some compile errors from incorrectly resolved merge conflict in yablocko

Fix clippy lints in kesha, scalez-stor-client

Fix a benchmark compile error and pin criterion to 0.3.3

The newly released 0.3.4 marks some APIs as deprecated which breaks
builds that treat warnings as errors

Remove unused yabloko deps

Fix failing tests caused by more aggressive new version of quickcheck

* Increase timeout on repl failure tests to 120

This is a shot in the dark.  Trying to see if it reduces the flakiness
of this test

* Fix hanging rangy tests

Caused by the new quickcheck which produces more diabolical test data
values which surfaced an existing bug in the tests

* Replace hard-coded timeout with retry

When this test runs on CI under code coverage it periodically fails due
to the hard-coded timeout.  Now it will retry until the server starts

* Fix flaky vault integration test better

Last time I fixed it I missed a second place where we need to wait for
the server to start.  Now it's done properly.

* Add more asserts around read path

The `ingest_long_chain` integration test on Blobs service failed on CI
with a troubling error about a read not returning all of the requested
data.  I don't know the cause, but in this change additional diagnostic
asserts are added to help narrow it down when it happens again.

* Stop running coverage tests with excessive threads

While this does a good job of breaking code under test, it also kills
performance and makes the tests flakier.