Add new tests and a ugly workaround for crash bug in merge operator (#1787)
* Add new tests and a ugly workaround for crash bug in merge operator

Summary
===
It turns out that the crash in the merge operator for `chunk_locations`
was due to an incredibly odd behavior in RocksDB in which the fully
merged value for a record is stored as a merge record and therefore
passed to the merge operator sometimes (the test that reproduces this
does so by closing the database and re-openning it).

Once I figured out the steps to reproduce and implemented a failing
test, it took a long time to figure out what was going on because it
initially just looked like the merge operands were corrupted.

The fix for this is pretty ugly: the scalez-kv merge operator code is
modified to wrap merge operands in `MergeOperandHack`, an aptly named
private enum that provides us a way to look at the serialized merge
operand passed to the merge operator and determine if it is in fact a
merge operand, or a fullly merged value.  It does this by appending a
distinct signature to the end of all serialized merge operands.  That's
a really ugly hack, but merge operands are not exposed anywhere in the
Rocks API except within the merge operator so we can tamper with the
serialized format however we want, whereas the fully merged value is
also returned from operations like `Get` so we don't have the same
ability to wrap it in some envelop.

This means that if a fully merged value, when serialized, ever ends with
the same byte string used as a signature for merge operands, then the
merge will fail with an error.  Hopefully this is not a very probable
occurence.

Other Changes
===

There were some changes I made during the initial troubleshooting phase
when I was still trying to locate the cause of the crash.  They are not
actally part of the fix, however they're useful so I'm keeping them:

* Fix panic message when slice is the wrong size

  The previous message wasn't being expanded into a string literal
  correctly so it was very confusing and ugly.

* Update to the latest RocksDB 6.15.4

  I had hoped that the strange behavior with the merge operator was a bug
  in Rocks and would be fixed in the latest version.  Sadly, no such luck,
  it's still happening.

  I'm sticking with the new version anyway, since I've gone to the trouble
  of the update.

* Fix cargo deny errors and resolved some issues exposed by updating

Fixing the cargo deny issues involved updating the `darling` dep from
0.10 to 0.12.  In 0.12 a bug was fixed in which Darling was previously
silently ignoring certain malformed attribute values on proc macros,
and now it fails loudly instead.  There were a few places in the
matroskin and cheburashka code bases that then broke, and required
some effort to fix.

* Disable Debian 8 builds

We decided not long ago that, since Debian 8 is already EOL, we wouldnt'
spend any more effort to keep those builds running; as soon as we run
into a Debian 8 problem that requires work to resolve, we'll stop
supporting Debian 8.

That moment has arrived.  The latest RocksDB code in this PR uses a
C++14 feature that doesn't compile right on GCC 4.9, which is what ships
with Debian 8.  So here we are.