Introduce a scalez-stor DB ID that truly doesn't ever change (#1888)
When introducing the concept of a scalez-store database ID (defined as
the RocksDB database ID of the blobs database) it was based on the
assumption that RocksDB IDs never change.  However it turns out that
when taking a database checkpoint, Rocks generates a new ID for the
checkpointed version of the database.  Therefore, when we load a shard
from an S3 snapshot (which comes from a checkpoint), the RocksDB
database ID is that of the checkpoint the snapshot came from, not the
original database.

This leads to a lot of problems, not least being defect #1883.

The fix was, unfortunately, not straightforward.  There's no way to
change the behavior of Rocks, nor is there a way to override the ID
assigned to a database.

Instead, I did something that I've been meaning to do for a while but
until now had no strong reason to: introduce the concept of KV
store-level named properties, where each property value can be of an
arbitrary type.  The first (and so far only) use for this mechanism
is...you guessed it...a new `KeyValueStoreId` which is a ULID and which
really is unique and really doesn't change when loading from a
checkpoint.

This also means our ID type is an actual binary ID, not a string that we
have to parse into a UUID in upper layers.  That's more efficient and
also has better type safety.

Unfortunately this is a breaking change, as the new ID type is a ULID
and the old ID type was a UUID, and the new ULID is obviously not going
to be backwards compatible with the old UUID.

I've renamed what used to be called `db_id` to `rocks_db_id`, and
introduced a new field `store_id` which is the `KeyValueStoreId`.  I've
kept the RocksDB ID in all spans and log events where it was previously,
since this is also useful information.  Not only does Rocks do its
logging in terms of the RocksDB ID (obviously it doesn't know about
higher-level ID concepts), but since the ID changes every time we load
from a checkpoint, this ID will also serve as something like a session
ID, so every time scalez-stor loads a shard from a snapshot and serves
it this ID will change.

Closes #1883