Add telemetry exfil to lambdas (#2119)
Summary
===
This relies on the new 2.0 version of `cheburashka` with substantial
changes to the config API, as a result this PR contains a lot of rather
trivial changes to some unrelated parts of the code base to bring the
rest of the code up to date.

The bulk of the real work here is in `nufnuf` and
`async_aws_lambda_runtime_core`.  Here, the new `cheburashka` features
allow us to extract individual JSON log events as they are logged, and
buffer them in the `TelemetryBuffer` just like we do for metrics today.
Then both metrics and logs are flushed to SQS after the lambda
invocation is completed.

This is still far from ideal.  Using the old, sync version of the Lambda
runtime means that we do too much of this logic within the context of a
single AWS invocation when really it should be something done
asynchronously in the background of the runtime.  Nonetheless, this gets
us the capability we wanted: automatic exfil of all log events and
metrics for any lamda that uses `async_aws_lambda_runtime_core`.

cc @Andry75 be advised there is a small change in the Protobuf schema
for telemetry here which the blue stack will need to be aware of and
handle.

Closes #1790