Fix backup failures after creating vault over existing one. (#2220)
Closes #2078 

This PR introduces `FileHasher` trait, responsible for reading file and generating checksum for it.

Now `smart_upload` function can be used with optional argument `file_hasher`.
If `file_hasher` is `None`, `smart_upload` behavior remains the same - it is assumed that files are never modified and it is enough to compare file name and size to decide whether we need to re-upload it to the object storage.

If `file_hasher` is not `None`, it will be used to compare checksums of local files with the checksums of the corresponding files in object storage.
There are two implementations of `FileHasher` trait:
* General-purpose `Crc32FileHasher`, which uses `crc32fast` crate.
* `SstFileHasher`, which is optimized for rocksDB SST files.

RocksDB has builtin capabilities of generating checksum for SST files on their creation (using CRC32c algorithm by default). This PR enables it. `SstFileHasher` then retrieves checksums from `LiveFileMetadata`. This way we don't need to generate checksum for each file on each checkpoint - they are already generated by RocksDB and stored in manifest.

There is `FileChecksumGenCrc32c` struct, which is a wrapper for corresponding RocksDB C++ class. It is used by `SstFileHasher` if checkpoint is not available - when we download store from S3 and need to validate checksums.