Add detailed metrics to sync and snapshot operations (#2332)
Summary
===
Before this change, there were no metrics on the snapshot or sync
operations at all.  We could see RocksDB metrics about the `commit` and
`checkpoint` operations, and sort-of see where sync ops were happening
as a result, but we had no visibility into how long those operations
take, how much work is being done, what files are new vs already synced,
etc.

This changes that in some key ways.  Details below.

Closes #942

Details
===

Refactor
---
Most of the change is in `scalez-kv`, specifically the `objstore`
module.  Before this there was a single Rust source file for all of
this, including syncing, snapshots, job queueing, etc.  This became
unweildy so I factored them out into submodules `sync`, `snapshot`,
`progress`, `hasher`, and the `objstore` mod itself.  All of this will
show up as changes in the diff, but the actual substantive changes are
just related to metrics gathering.

Labels
---
I added a `scalez_kv::labels` module which contains `cheburashka` label
structs for consistently applying labels to log events and metrics
related to KV stores.  It's really important to be consistent in the use
of labels, particularly on related metrics, and this provides us with a
type-safe way to do that.  As a result you'll see labels structs being
passed around in a few places.

Store Types
---
I wanted to be able to group metrics by which scalez-stor service they
related to (`chunks`, `blobs`, etc).  Before this was kind-of possible
using the `db-path` label, by dumb luck, because our DB paths are always
something like `/mnt/elastio/chunks` or something, but there's no
guarantee that will always be the case.

So I added `KeyValueStore::store_type` which is a `&'static str` which
defaults to the Rust type name of the impl.  This was also added as a
label and a log event field in a bunch of places.  It'll make present
and future work with dashboarding a lot easier.

Progress Reporter Metrics
---

The `SyncProgressReporter` trait in `domovoi` was originally created to
provide a convenient place to capture important sync metrics, but the
metrics themselves were not implemented.  Now they are, so we can see
detailed metrics about the files and bytes that are synced, including
new bytes that are uploaded and skipped bytes.  There's also a histogram
for tracking approximate byte size of files.

Sync Metrics
---
The sync operation itself is a bit higher level than file-level progress
reporting, but is probably the most important in understanding the
behavior of the sync operation.

Snapshot Metrics
---
Since a snapshot is just a sync with some additional database and
metadata files added to object storage to make a complete database
backup, most of the sync metrics also apply.  However some additional
metrics for counting snapshots and latency are needed because of this
additional logic that could also introduce strange performance contours.