Fix zstd dep so static lib builds are supported with both lz4 and zstd
The lz4 and zstd C codebases both have a copy-paste of the same XXHash
code (typical C/C++ discipline).  When compiling a static lib target,
`roxide` was failing to link because of the symbol collission.

Thankfully the `zstd` authors thought of this and provide a preprocessor
definition that renames the XXHash symbols with a prefix.  I just had to
enable it.