Introduce `Cache` type and support for explicit caches
Before this change, `roxide` had no direct concept of block caches; only
the options map that can contain a `block_cache` option which we set by
default to `1G`.  The problem with that is that it means our ability to
control which CFs use which cache is limited, and it's completely
impossible to share a single cache between multiple RocksDB databases in
a single process.

To address this I added a new module `cache` and struct `Cache`, which
wraps the RocksDB C++ `Cache` object.  This can even be deserialized
from a string or integer using the same config syntax that Rocks uses,
which will come in handy a bit later when I integrate these changes into
`scalez-kv`.

This also changed the `DbOptions` config logic a great deal, because we
now allow either an explicit `Cache` object to be set for the whole
database or a specific CF, or the existing options map-based config
which was used before.  These can even be mixed and matched, so now
higher-level code like `scalez-kv` and `scalez-stor` will have a lot
more control over caching, including allowing `scalez-stor` to build one
single block cache for all three databases to use.

Testing this proved to be a real pain in the ass, because the RocksDB
C++ API does not expose the block cache that was used via the public
API, so writing a test that verifies all CFs are using the same block
cache and not unique ones was very difficult.  I ended up having to do
some pretty nasty hacks, mostly in `db_options.h` and `db_options.rs`,
to make this work.  Those hacks required some changes to how the RocksDB
code is compiled in `roxide-librocksdb-sys/build.rs`, although those
changes themselves aren't particularly hacky.

This is part of the fix for elastio/elastio#3062, but it doesn't close
it.  That will require further changes to scalez as well.