Replace DynTraitBox with DynTraitArc
This helper struct was used to smuggle boxed Rust trait objects into C++
and then back out again.  It used `Box<T>` which worked fine but is not
clonable.  We need `DbOptions` to be clonable to support a convenience
method to open a checkpoint with the same settings as the live DB
(critically, the merge operators) to discover the files in that
checkpoint, and that's not possible unless it contains only clonable
types.