Move calculating seq num and SST files to `checkpoint` op
This required a lot of changes but I think it'll be worth it.

Database objects now have a reference to the database's default block
cache and a clone of the `DbOptions` used to open the database.  Now
when taking a checkpoint of the database, that checkpoint also gets a
copy of the DbOptions and the cache.  This allows the `Checkpoint`
struct to have a `get_files` method which opens the checkpoint database
read-only with some additional config options to make the open happen
as fast as possible, and uses the live database's block cache rather
than letting Rocks allocated a separate cache for no reason.

This also uses the C++ checkpoint API, and exposes both the checkpoint
seq number and the `log_size_for_flush` parameter in the Rust API so
scalez can use it.

Not only will this make taking checkpoints for scalez snapshots a bit
faster because of the optimized DB open, but it should also resolve the
problem of rapid growth in the number of metrics timeseries because each
time the checkpoint was opened it was logging database metrics for that
checkpoint as if it were a unique database.