Introduce `Cache` type and support for explicit caches plus checkpoint improvements (#5)
Summary
===

This change adds Rust-level types for creating and managing block caches at the DB and CF level.

It also includes some adjacent target-of-opportunity changes described in more detail below.

This PR is part of the fix for elastio/elastio#3062, but it doesn't close
it.  That will require further changes to scalez as well.

Cache Support
---

Before this change, `roxide` had no direct concept of block caches; only the options map that can contain a
`block_cache` option which we set by default to `1G`.  The problem with that is that it means our ability to control
which CFs use which cache is limited, and it's completely impossible to share a single cache between multiple RocksDB
databases in a single process.

To address this I added a new module `cache` and struct `Cache`, which wraps the RocksDB C++ `Cache` object.  This can
even be deserialized from a string or integer using the same config syntax that Rocks uses, which will come in handy a
bit later when I integrate these changes into `scalez-kv`.

This also changed the `DbOptions` config logic a great deal, because we now allow either an explicit `Cache` object to
be set for the whole database or a specific CF, or the existing options map-based config which was used before.  These
can even be mixed and matched, so now higher-level code like `scalez-kv` and `scalez-stor` will have a lot more control
over caching, including allowing `scalez-stor` to build one single block cache for all three databases to use.

Testing this proved to be a real pain in the ass, because the RocksDB C++ API does not expose the block cache that was
used via the public API, so writing a test that verifies all CFs are using the same block cache and not unique ones was
very difficult.  I ended up having to do some pretty nasty hacks, mostly in `db_options.h` and `db_options.rs`, to make
this work.  Those hacks required some changes to how the RocksDB code is compiled in `roxide-librocksdb-sys/build.rs`,
although those changes themselves aren't particularly hacky.

Checkpoint Improvements
---

This also modifies the `checkpoint` API to expose the `log_size_for_flush` parameter which was previously hard-coded to
zero, so scalez can now control whether or not all CFs are flushed on every checkpoint.

There's also an optimization here which is intended to help with scalez performance, which is that the `Checkpoint`
struct now has a method `get_files` which does the same thing as the DB-level `get_live_files`, namely it gets a list of
all active SST files and their metadata, but this implementation is more optimized:

* It re-uses the source database's block cache so no additional cache needs to be allocated
* It sets some DB-level options that the Rocks wiki suggests improve the speed with with a database is opened
* It does this with the C++ API directly, without instantiating a Rust DB wrapper.  I suspect this will help a lot to
	reduce the excessive metrics described in elastio/elastio#2457

Other Changes
---

In this PR I also expose additional fields in `SstFile`, including the range of sequence numbers and keys in every SST
file.  We have some ideas about some intelligent manual compaction in scalez that would benefit from having access to
this information, and since I'm already making a breaking change I decided to combine all breaking changes into one.

I also updated `rust-toolchain` from 1.51 to 1.54, which required making some minor edits to address code that now
causes warnings but didn't in 1.51 and before.

I had to make `DbOptions` struct clonable, which probably should have been done a long time ago.  This required changing
the `DynTraitBox` helper struct into `DynTraitArc`, so that things like custom merge operators and loggers are always
clonable.