Add `jemalloc` feature to `roxide`
This new feature, not enabled by default, will compile `jemalloc` from
source using the `jemallocator` project, and compile RocksDB with
jemalloc support activated.  RocksDB issues and our own tests show that
jemalloc is a much better fit for how RocksDB manages memory.

Note this does not cause the Rust code to use jemalloc; that must be set
in the `main.rs` for the Rust binary.  This causes RocksDB to use
jemalloc for its memory management.