Add `liburing` submodule and build steps
I thought it made more sense to compiled rocksdb against the liburing OS
package on the CI runner, but there were two problems:

- the liburing package on Ubuntu 22.04 is too old to compile the latest
  RocksDB
- The liburing docs explicitly say there's no need to try to synchronize
  the liburing version with the corresponding kernel version

That, plus the fact that the Ubuntu OS package doesn't install a static
lib so we would have produced executables with a runtime dep on the
liburing lib, made it a no-brainer to add a submodule with the latest
released `liburing` and build it from source if the `io_uring` feature
is enabled.  Since it's just a few C files it was trivial to get it
working.