Update RocksDB (and other deps) (#28)
* Update RocksDB to 8.0.0, and also a few Rust deps

This is a breaking change as Rocks 8 removed some config options that
were part of the public API.

I also took this opportunity to update to Rust 1.65, and to update the
Rust dependencies to the latest versions.

* Update native deps:

lz4 - v1.9.4
zstd - v1.5.4
snappy - v1.1.10

* Fix some build errors and test failures after updating to latest

* Suppress snappy compile warning

* Fix rust formatting

* Fix cargo deny issues

* Fix lints

* Modify build to add support for more advanced Rocks capabilities.

This adds the `io_uring` feature to `roxide`, which if enabled will
compile RocksDB with io_uring support.  The resulting Rust crate will
have a runtime dependency on the `liburing` shared library.

* Get async support working (I think)

* Try to get build working with folly and io_uring

* Fix formatting

* Set credentials thew new way

* Fix folly checkout

* DOn't install gflags package

* Try some more bullshit to get folly working

* Try desperate shit

* Try to fix folly build again

* Maybe fix toolchain install logic

* Fix clippy lint and folly build

* Update available RocksDB metrics

* Try (and fail) to get `folly` to compile reliably

* Remove `folly` submodule

Something got fucked up so I need to re-create it again

* Re-create `folly` submodule

* Fix cargo deny issues

* Update to RocksDB 8.1.1 (from 8.0.0)

* Comment out the `folly` feature to `--all-features` works

* Add a note about async support

* Add `liburing` submodule and build steps

I thought it made more sense to compiled rocksdb against the liburing OS
package on the CI runner, but there were two problems:

- the liburing package on Ubuntu 22.04 is too old to compile the latest
  RocksDB
- The liburing docs explicitly say there's no need to try to synchronize
  the liburing version with the corresponding kernel version

That, plus the fact that the Ubuntu OS package doesn't install a static
lib so we would have produced executables with a runtime dep on the
liburing lib, made it a no-brainer to add a submodule with the latest
released `liburing` and build it from source if the `io_uring` feature
is enabled.  Since it's just a few C files it was trivial to get it
working.