Make roxide compile on newer versions of rust compiler.
I tried to use [`rust-marker`](https://github.com/rust-marker/marker) in elastio/elastio repo to write a custom "clippy" lint. However, it depends on a specific newer nightly compiler versions for compilation which no longer accepts the code in `roxide`.

The problem is that the `Send` autotrait impl was disabled for `OpenKeyRange` type. I implemented that trait for all `T` in `OpenKeyRange<T>`. To work around this I introduced a newtype over the `(*mut u8, usize)` and implemented `Send` for it.